feat: image placeholders
