+++ components/project-list.tsx
@@ -28,7 +28,7 @@ const ProjectList = ({ projects }: ProjectListProps) => {
                         className="block mt-9 transition-all w-full h-[256px] relative overflow-hidden rounded-0 focus:outline-none ring-gray-900 focus:ring-4 hover:ring-4 ring-opacity-80 sm:flex-shrink-0 sm:w-1/2 sm:h-[250px] sm:mt-0 sm:rounded-tl-[16px] sm:rounded-bl-[16px] md:rounded-2xl"
                         aria-label={`${project.message} source`}
                     >
-                        <Image src={project.image} alt={project.title} layout="fill" objectFit="cover" />
+                        <Image src={project.image} blurDataURL={project.blurDataURL} alt={project.title} placeholder="blur" layout="fill" objectFit="cover" />
                         <ImageShadow className="sm:rounded-tl-[16px] sm:rounded-bl-[16px] md:rounded-2xl" />
                     </a>
                 </li>
