+++ package.json
@@ -17,9 +17,11 @@
         "lodash": "^4.17.21",
         "markdown-to-jsx": "^7.1.3",
         "next": "11.0.1",
+        "plaiceholder": "^2.0.1",
         "react": "17.0.2",
         "react-dom": "17.0.2",
-        "react-use": "^17.2.4"
+        "react-use": "^17.2.4",
+        "sharp": "^0.28.3"
     },
     "devDependencies": {
         "@types/lodash": "^4.14.171",
