+++ pages/index.tsx
@@ -3,7 +3,7 @@ import IntroSection from "../components/intro-section";
 import AboutSection from "../components/about-section";
 import ExperienceSection from "../components/experience-section";
 import ContactSection from "../components/contact-section";
-import { getProjects } from "../utils/markdown";
+import { getProjects } from "../utils/data";
 import { GetStaticProps } from "next";
 import { Project } from "../utils/types";
