+++ utils/types.ts
@@ -1,6 +1,7 @@
 export interface Project {
     title: string;
     image: string;
+    blurDataURL: string;
     source: string;
     message: string;
     labels: string[];
